requirements.txt
requests
beautifulsoup4
urllib3
python-telegram-bot==20.7